Add files via upload
WeatherNow is a website that tells you what the weather is like at the moment in different cities. It's handy if you want to know if it's sunny, rainy, or something else in a specific place. It focuses on the current weather conditions rather than giving forecasts for the future. So, if you're wondering what the weather is like in Paris or New York right now, WeatherNow has got you covered!